Remove duplicated WrappedRcRefCell and add macro to create wrapped types
