+++ crates/hyperqueue/src/common/mod.rs
@@ -12,10 +12,9 @@ pub mod serverdir;
 pub mod setup;
 pub mod strutils;
 pub mod timeutils;
-pub mod wrapped;
 
 use tako::common::macros::define_id_type;
 
 define_id_type!(JobId);
 
-pub use wrapped::WrappedRcRefCell;
+pub use tako::common::WrappedRcRefCell;
